+++ CHANGES.md
@@ -6,6 +6,9 @@
   - Decoupled `--log-level` from command-line logging. Changed default log-level to
     ERROR and made it apply to automatically generated log files as well.
 
+### Fixed
+  - Fixed the pipeline failing on `jre_options` (now `jre-option`) in config files.
+
 
 ## [1.3.1] - 2020-09-01