+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [1.3.2] - 2020-09-03
 
 ### Added
   - Added `-v` and `--version` to all command-line tools.
@@ -720,7 +720,8 @@ the (partially) updated documentation now hosted on ReadTheDocs.
   - Switching to more traditional version-number tracking.
 
 
-[Unreleased]: https://github.com/MikkelSchubert/paleomix/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/MikkelSchubert/paleomix/compare/v1.3.2...HEAD
+[1.3.2]: https://github.com/MikkelSchubert/paleomix/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/MikkelSchubert/paleomix/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/MikkelSchubert/paleomix/compare/v1.2.14...v1.3.0
 [1.2.14]: https://github.com/MikkelSchubert/paleomix/compare/v1.2.13.8...v1.2.14